Fix tare bug with raw readings. (#3)
Zeroing the scale now uses the raw value. Previously, `tare()` used
the `read()` function directly, which accounts for the offset. Calling
`tare()` multiple times would result in the weight not being reset.

Adjust the read interface to return a `NotReadyError` when trying to
read from the device before it has a value ready. Previously, it
woud return the minimum mass possible, which was hacky.

Co-authored-by: DaneSlattery <dane_s@umantec.net>